refactor: move controller to loader
