+++ lib/base_loader.js
@@ -178,21 +178,6 @@ class EggLoader {
   }
 
   /**
-   * 加载 app/controller 目录下的文件
-   *
-   * @param {Object} opt - loading 参数
-   */
-  loadController(opt) {
-    const app = this.app;
-    opt = Object.assign({ lowercaseFirst: true }, opt);
-    const controllerBase = path.join(this.options.baseDir, 'app/controller');
-
-    this.loadToApp(controllerBase, 'controller', opt);
-    app.controllers = app.controller;
-    app.coreLogger.info('[egg:loader] Controller loaded: %s', controllerBase);
-  }
-
-  /**
    * 返回 egg 需要加载的目录
    *
    * 1. 核心框架目录，目录为框架根目录下的 lib/core 目录，框架根目录来自 {@link EggLoader#eggPaths}
@@ -258,6 +243,7 @@ const loaders = [
   require('./proxy_loader'),
   require('./service_loader'),
   require('./loader/middleware'),
+  require('./loader/controller'),
 ];
 
 for (const loader of loaders) {
